add block manager (merge request !21)

+++ src/numerous_llm/block_manager/memory_block.h
@@ -2,6 +2,7 @@
 
 ==============================================================================*/
 #pragma once
+#include <cstdint>
 
 namespace numerous_llm {
 
@@ -23,13 +24,13 @@ enum MemoryDevice {
 // The memory block information.
 struct MemoryBlock {
   // block id, unique in global.
-  int block_id;
+  int64_t block_id;
 
   // block size, in bytes.
-  int block_size;
+  int64_t block_size;
 
   // The reference count of current block.
-  int ref_count = 0;
+  int64_t ref_count = 0;
 
   // /The device of this block, CPU or GPU or NPU.
   MemoryDevice device;
